# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LightRAG Server API",
        "description": "Providing API for LightRAG core, Web UI and Ollama Model Emulation\n\n[View ReDoc documentation](/redoc)",
        "version": "0273"
    },
    "paths": {
...
        "/query": {
            "post": {
                "tags": [
                    "query"
                ],
                "summary": "Query Text",
                "description": "Comprehensive RAG query endpoint with non-streaming response. Parameter \"stream\" is ignored.\n\nThis endpoint performs Retrieval-Augmented Generation (RAG) queries using various modes\nto provide intelligent responses based on your knowledge base.\n\n**Query Modes:**\n- **local**: Focuses on specific entities and their direct relationships\n- **global**: Analyzes broader patterns and relationships across the knowledge graph\n- **hybrid**: Combines local and global approaches for comprehensive results\n- **naive**: Simple vector similarity search without knowledge graph\n- **mix**: Integrates knowledge graph retrieval with vector search (recommended)\n- **bypass**: Direct LLM query without knowledge retrieval\n\nconversation_history parameteris sent to LLM only, does not affect retrieval results.\n\n**Usage Examples:**\n\nBasic query:\n```json\n{\n    \"query\": \"What is machine learning?\",\n    \"mode\": \"mix\"\n}\n```\n\nBypass initial LLM call by providing high-level and low-level keywords:\n```json\n{\n    \"query\": \"What is Retrieval-Augmented-Generation?\",\n    \"hl_keywords\": [\"machine learning\", \"information retrieval\", \"natural language processing\"],\n    \"ll_keywords\": [\"retrieval augmented generation\", \"RAG\", \"knowledge base\"],\n    \"mode\": \"mix\"\n}\n```\n\nAdvanced query with references:\n```json\n{\n    \"query\": \"Explain neural networks\",\n    \"mode\": \"hybrid\",\n    \"include_references\": true,\n    \"response_type\": \"Multiple Paragraphs\",\n    \"top_k\": 10\n}\n```\n\nConversation with history:\n```json\n{\n    \"query\": \"Can you give me more details?\",\n    \"conversation_history\": [\n        {\"role\": \"user\", \"content\": \"What is AI?\"},\n        {\"role\": \"assistant\", \"content\": \"AI is artificial intelligence...\"}\n    ]\n}\n```\n\nArgs:\n    request (QueryRequest): The request object containing query parameters:\n        - **query**: The question or prompt to process (min 3 characters)\n        - **mode**: Query strategy - \"mix\" recommended for best results\n        - **include_references**: Whether to include source citations\n        - **response_type**: Format preference (e.g., \"Multiple Paragraphs\")\n        - **top_k**: Number of top entities/relations to retrieve\n        - **conversation_history**: Previous dialogue context\n        - **max_total_tokens**: Token budget for the entire response\n\nReturns:\n    QueryResponse: JSON response containing:\n        - **response**: The generated answer to your query\n        - **references**: Source citations (if include_references=True)\n\nRaises:\n    HTTPException:\n        - 400: Invalid input parameters (e.g., query too short)\n        - 500: Internal processing error (e.g., LLM service unavailable)",
                "operationId": "query_text_query_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "api_key_header_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Api Key Header Value"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful RAG query response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueryResponse",
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "string",
                                            "description": "The generated response from the RAG system"
                                        },
                                        "references": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "reference_id": {
                                                        "type": "string"
                                                    },
                                                    "file_path": {
                                                        "type": "string"
                                                    },
                                                    "content": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "description": "List of chunk contents from this file (only included when include_chunk_content=True)"
                                                    }
                                                }
                                            },
                                            "description": "Reference list (only included when include_references=True)"
                                        }
                                    },
                                    "required": [
                                        "response"
                                    ]
                                },
                                "examples": {
                                    "with_references": {
                                        "summary": "Response with references",
                                        "description": "Example response when include_references=True",
                                        "value": {
                                            "response": "Artificial Intelligence (AI) is a branch of computer science that aims to create intelligent machines capable of performing tasks that typically require human intelligence, such as learning, reasoning, and problem-solving.",
                                            "references": [
                                                {
                                                    "reference_id": "1",
                                                    "file_path": "/documents/ai_overview.pdf"
                                                },
                                                {
                                                    "reference_id": "2",
                                                    "file_path": "/documents/machine_learning.txt"
                                                }
                                            ]
                                        }
                                    },
                                    "with_chunk_content": {
                                        "summary": "Response with chunk content",
                                        "description": "Example response when include_references=True and include_chunk_content=True. Note: content is an array of chunks from the same file.",
                                        "value": {
                                            "response": "Artificial Intelligence (AI) is a branch of computer science that aims to create intelligent machines capable of performing tasks that typically require human intelligence, such as learning, reasoning, and problem-solving.",
                                            "references": [
                                                {
                                                    "reference_id": "1",
                                                    "file_path": "/documents/ai_overview.pdf",
                                                    "content": [
                                                        "Artificial Intelligence (AI) represents a transformative field in computer science focused on creating systems that can perform tasks requiring human-like intelligence. These tasks include learning from experience, understanding natural language, recognizing patterns, and making decisions.",
                                                        "AI systems can be categorized into narrow AI, which is designed for specific tasks, and general AI, which aims to match human cognitive abilities across a wide range of domains."
                                                    ]
                                                },
                                                {
                                                    "reference_id": "2",
                                                    "file_path": "/documents/machine_learning.txt",
                                                    "content": [
                                                        "Machine learning is a subset of AI that enables computers to learn and improve from experience without being explicitly programmed. It focuses on the development of algorithms that can access data and use it to learn for themselves."
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "without_references": {
                                        "summary": "Response without references",
                                        "description": "Example response when include_references=False",
                                        "value": {
                                            "response": "Artificial Intelligence (AI) is a branch of computer science that aims to create intelligent machines capable of performing tasks that typically require human intelligence, such as learning, reasoning, and problem-solving."
                                        }
                                    },
                                    "different_modes": {
                                        "summary": "Different query modes",
                                        "description": "Examples of responses from different query modes",
                                        "value": {
                                            "local_mode": "Focuses on specific entities and their relationships",
                                            "global_mode": "Provides broader context from relationship patterns",
                                            "hybrid_mode": "Combines local and global approaches",
                                            "naive_mode": "Simple vector similarity search",
                                            "mix_mode": "Integrates knowledge graph and vector retrieval"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "detail": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "detail": "Query text must be at least 3 characters long"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Query processing failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "detail": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "example": {
                                    "detail": "Failed to process query: LLM service unavailable"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "QueryRequest": {
                "properties": {
                    "query": {
                        "type": "string",
                        "minLength": 3,
                        "title": "Query",
                        "description": "The query text"
                    },
                    "mode": {
                        "type": "string",
                        "enum": [
                            "local",
                            "global",
                            "hybrid",
                            "naive",
                            "mix",
                            "bypass"
                        ],
                        "title": "Mode",
                        "description": "Query mode",
                        "default": "mix"
                    },
                    "only_need_context": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Only Need Context",
                        "description": "If True, only returns the retrieved context without generating a response."
                    },
                    "only_need_prompt": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Only Need Prompt",
                        "description": "If True, only returns the generated prompt without producing a response."
                    },
                    "response_type": {
                        "anyOf": [
                            {
                                "type": "string",
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Response Type",
                        "description": "Defines the response format. Examples: 'Multiple Paragraphs', 'Single Paragraph', 'Bullet Points'."
                    },
                    "top_k": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Top K",
                        "description": "Number of top items to retrieve. Represents entities in 'local' mode and relationships in 'global' mode."
                    },
                    "chunk_top_k": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chunk Top K",
                        "description": "Number of text chunks to retrieve initially from vector search and keep after reranking."
                    },
                    "max_entity_tokens": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Entity Tokens",
                        "description": "Maximum number of tokens allocated for entity context in unified token control system."
                    },
                    "max_relation_tokens": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Relation Tokens",
                        "description": "Maximum number of tokens allocated for relationship context in unified token control system."
                    },
                    "max_total_tokens": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "minimum": 1.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Total Tokens",
                        "description": "Maximum total tokens budget for the entire query context (entities + relations + chunks + system prompt)."
                    },
                    "hl_keywords": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Hl Keywords",
                        "description": "List of high-level keywords to prioritize in retrieval. Leave empty to use the LLM to generate the keywords."
                    },
                    "ll_keywords": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Ll Keywords",
                        "description": "List of low-level keywords to refine retrieval focus. Leave empty to use the LLM to generate the keywords."
                    },
                    "conversation_history": {
                        "anyOf": [
                            {
                                "items": {
                                    "additionalProperties": true,
                                    "type": "object"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conversation History",
                        "description": "History messages are only sent to LLM for context, not used for retrieval. Format: [{'role': 'user/assistant', 'content': 'message'}]."
                    },
                    "user_prompt": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Prompt",
                        "description": "User-provided prompt for the query. If provided, this will be used instead of the default value from prompt template."
                    },
                    "enable_rerank": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Enable Rerank",
                        "description": "Enable reranking for retrieved text chunks. If True but no rerank model is configured, a warning will be issued. Default is True."
                    },
                    "include_references": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Include References",
                        "description": "If True, includes reference list in responses. Affects /query and /query/stream endpoints. /query/data always includes references.",
                        "default": true
                    },
                    "include_chunk_content": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Include Chunk Content",
                        "description": "If True, includes actual chunk text content in references. Only applies when include_references=True. Useful for evaluation and debugging.",
                        "default": false
                    },
                    "stream": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stream",
                        "description": "If True, enables streaming output for real-time responses. Only affects /query/stream endpoint.",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "query"
                ],
                "title": "QueryRequest"
            },
            "QueryResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response",
                        "description": "The generated response"
                    },
                    "references": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/ReferenceItem"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "References",
                        "description": "Reference list (Disabled when include_references=False, /query/data always includes references.)"
                    }
                },
                "type": "object",
                "required": [
                    "response"
                ],
                "title": "QueryResponse"
            },
            "ReferenceItem": {
                "properties": {
                    "reference_id": {
                        "type": "string",
                        "title": "Reference Id",
                        "description": "Unique reference identifier"
                    },
                    "file_path": {
                        "type": "string",
                        "title": "File Path",
                        "description": "Path to the source file"
                    },
                    "content": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content",
                        "description": "List of chunk contents from this file (only present when include_chunk_content=True)"
                    }
                },
                "type": "object",
                "required": [
                    "reference_id",
                    "file_path"
                ],
                "title": "ReferenceItem",
                "description": "A single reference item in query responses."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "description": "OAuth2 Password Authentication",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "login"
                    }
                }
            }
        }
    }
}